mysite/requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
django-taggit==5.0.1
Markdown==3.5
psycopg2-binary==2.9.9